# api.yaml
swagger: "2.0"
info:
  title: Viviendas reales
  description: descripción de la API.
  version: 1.0.0
host: localhost:5000
basePath: /api
schemes:
  - http
  - https
paths:
...
  /housing:
    post:
      responses:
        201:
          description: recurso creado
        400:
          description: faltan atributos requeridos
      tags:
        - housing
      description: crea una nueva vivienda
      parameters:
        - name: housing
          in: body
          required: true
          schema:
            type: object
            properties:
              address:
                type: object
                properties:
                  state:
                    type: string
                  province:
                    type: string
                  road:
                    type: string
                    enum:
                      - street
                      - avenue
                      - road
                      - square
                  name:
                    type: string
                  number:
                    type: string
                  floor:
                    type: string
                  ladder:
                    type: string
              description:
                type: string
              price:
                type: number
              surface:
                type: number
              owner:
                type: string
    get:
      responses:
        200:
          description: success
        404:
          description: no existe el documento en base de datos
      tags:
        - housing
      description: obtiene el un conjunto de recursos paginados
      parameters:
        - name: limit
          in: query
          description: limite de los recursos obtenidos
          type: number
        - name: page
          in: query
          type: number
          description: página actual
    put:
      responses:
        201:
          description: recurso actualizado correctamente
        404:
          description: recurso no encontrado
      tags:
        - housing
      description: actualiza la vivienda de manera absoluta
      parameters:
        - name: housing
          in: body
          required: true
          schema:
            type: object
            properties:
              id:
                type: string
              address:
                type: object
                properties:
                  state:
                    type: string
                  province:
                    type: string
                  road:
                    type: string
                    enum:
                      - street
                      - avenue
                      - road
                      - square
                  name:
                    type: string
                  number:
                    type: string
                  floor:
                    type: string
                  ladder:
                    type: string
              description:
                type: string
              price:
                type: number
              surface:
                type: number
              owner:
                type: string
  /housing/{id}:
    get:
      responses:
        200:
          description: recurso recuperado correctamente
        404:
          description: recurso no encontrado
      tags:
        - housing
      description: recupera el recurso housing coincidente con la id pasada por parámetro
      parameters:
        - name: id
          type: string
          required: true
          in: path
          description: id del recurso a recuperar
    delete:
      responses:
        203:
          description: recurso borrado correctamente
        404:
          description: recurso a borrar no existe o no se encuentra
      tags:
        - housing
      description: borra el recurso con id coincidente con la pasada por parámetro
      parameters:
        - name: id
          type: string
          in: path
          required: true
          description: id del recurso a recuperar
    patch:
      responses:
        201:
          description: recurso actualizado correctamente
        404:
          description: no existe o no se encuentra recurso
      tags:
        - housing
      description: actualización parcial del recurso housing con id coincidente con la pasada por parámetro
      parameters:
        - name: file
          required: true
          in: formData
          type: file
...
  /image:
    post:
      tags:
        - image
      responses:
        201:
          description: recurso creado correctamente
        400:
          description: falta algún campo requerido
      parameters:
        - name: image
          required: true
          in: body
          schema:
            type: object
            properties:
              title:
                type: string
              housing:
                type: string
        - name: file
          required: true
          in: formData
          type: file
    get:
      tags:
        - housing
      responses:
        200:
          description: recursos recuperados con éxito
        404:
          description: no existe la colección de datos
      parameters:
        - name: housing
          type: string
          in: path
          required: true
        - name: limit
          type: string
          in: query
          required: false
          description: límite de recursos recuperados en una petición
        - name: page
          type: string
          in: query
          required: false
          description: página actual
      description: recupera un conjunto de recursos de tipo image de manera paginada
  /image/:id:
    patch:
      tags:
        - image
      responses:
        201:
          description: recurso actualizado con éxito
        404:
          description: recurso no encontrado
      description: actualiza de manera parcial el recurso de tipo image que coincide con el parámetro id pasado por la 
                   query de la petición
      parameters:
        - name: id
          type: string
          in: path
          required: true
          description: id del recurso el cuál de pretende actualizar
        - name: file
          in: formData
          type: file
          required: true
          description: fichero nuevo
        - name: title
          in: formData
          type: string
          description: nuevo título
    delete:
      tags:
        - image
      responses:
        203:
          description: recurso borrado con éxito
        404:
          description: recurso no encontrado o no existe
...